+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## Bugfree.Spo.ExternalSharingCenter 1.0.2 (2016-08-04)
+
+* [#4](https://github.com/ronnieholm/Bugfree.Spo.ExternalSharingCenter/issues/4) - Ensure internal consistency between SharePoint lists on External Sharing Center
+
 ## Bugfree.Spo.ExternalSharingCenter 1.0.1 (2016-07-31)
 
 * [#3](https://github.com/ronnieholm/Bugfree.Spo.ExternalSharingCenter/issues/3) - External sharings with InvitedAs != AcceptedAs causes warning mail generation to fail